+++ lib/phaseout/phase_insert.c
@@ -9,9 +9,9 @@
 
 char *phase_insert(char *base, char add, char end)
 {
-    int base_length = phase_length(base, end);
+    size_t base_length = phase_length(base, end);
     char *array = phase_malloc((base_length + 1), end);
-    int i = 0;
+    size_t i = 0;
 
     for (; i < base_length; i++) {
         array[i] = base[i];
