+++ lib/phaseout/phase_malloc.c
@@ -7,11 +7,11 @@
 
 #include "PHASEOUT.h"
 
-char *phase_malloc(int size, char content)
+char *phase_malloc(size_t size, char content)
 {
     char *array = malloc(sizeof(char) * (size + 1));
 
-    for (int i = 0; i <= size; i++)
+    for (size_t i = 0; i <= size; i++)
         array[i] = content;
 
     return (array);
